Add QCRAM support to HPACK, part 3 - absolute indexing
Summary:
This adds absolute indexing.  The spec is somewhat underspecified here, but my best guess it that the base index is supposed to be like an absolute version of head_.  This won't work with table resizing, but as I understand it, that's dead in the current QUIC draft.

The encoder calls 'markBaseIndex' and serializes the index as an HPACK integer with 0 bit prefix (previously illegal in our code).  This sets the readBaseIndex for the encoder so that all serialized indecies are relative to the starting base index.

The decoder calls setBaseIndex when processing a block which handles both translations for indexed header decoding (readBaseIndex) and also potentially offset adds into the table (writeBaseIndex).

Reviewed By: dddmello

Differential Revision: D4943055

fbshipit-source-id: 0c33f16673305f721f55a122cb8030b301c2c649