CompressionSimulator
Summary:
This is a simulator to test the efficacy of the QCRAM/QPACK algorithms.  The input is a HAR file containing actual request start times and headers (such a file can be captured from a browser).  The inital set of knobs are RTT, loss rate, delay rate, max delay, and how many inital forced in-order packets to use.  You can also specify an RNG seed to compare how different algorithms fare with the same parameters.

Example

HPACK style (force receiver to decode in order)

  Stats:
  Blocks sent: 185
  Allowed OOO: 0
  Packet Losses: 14
  HOL Block Count: 9
  HOL Delay (ms): 1344
  Max Queue Buffer Bytes: 1342
  Uncompressed Bytes: 137265
  Compressed Bytes: 36216
  Compression Ratio: 73

QCRAM with same network behavior

  Stats:
  Blocks sent: 185
  Allowed OOO: 98
  Packet Losses: 14
  HOL Block Count: 7
  HOL Delay (ms): 1004
  Max Queue Buffer Bytes: 826
  Uncompressed Bytes: 137265
  Compressed Bytes: 42902
  Compression Ratio: 68

Reviewed By: dddmello

Differential Revision: D4971193

fbshipit-source-id: c2fcc53829f3d76418ad1b32b9554465eb7eacf8