Fix native upgrade with partial early response
Summary:
If a server begins responding to an upgrade request, but does not complete before the client's request completes, the server will attempt to send a second header block and fail a DCHECK in HTTP1xCodec about sending duplicate headers.

There was already a check for 'canSendHeaders' in HTTPDownstreamSession, but canSendHeaders was returning true in this case.  Tighten up the canSendHeaders logic

Reviewed By: dddmello

Differential Revision: D6329655

fbshipit-source-id: 3920b45d83c71a4632d5f00237922a849972cea5