Adding lastByteEvent tracker for sending body without EOM
Summary:
For most HTTP2 transactions, we send EOM when we send body data. However, for Whistle, when it wraps MQTT over HTTP2 transport, it does not send EOM after sending the body data. It keeps the transaction open for a long time and use sendBody and onBody to exchange messages.
Therefore, we need a separate flag (`trackLastByteFlushed`) to call lastByteFlushed() when we finish sending body data to the network.

Reviewed By: afrind

Differential Revision: D6252253

fbshipit-source-id: 981e38164df48617e2e06c7fef37050e46fc0e27