QPACK draft-5 updates
Summary:
Draft 5 makes a few changes over draft 1

1) inserts and deletes happen on the control stream
2) the serialized format has changed - static and dynamic table references now have separate namespaces identified by a bit in the instructions

To support inserts on the control stream, I refactored the encoder and decoder logic a bit to share code.

Note: draft-5 has a different scheme for deletes that I have not implemented yet.  The draft-1 scheme is sufficient for now because the simulator never resets any streams.

Reviewed By: dddmello

Differential Revision: D6441434

fbshipit-source-id: 33198d81018d1851feff5f0e109fceb03ea48e93