Fixing Header struct to always have the right code
Summary: It is possible to instantiate a Header instance with the wrong code.  For example the H2 and SPDY Codecs add headers in the production path with the wrong code.  The main use of the codes is during sorting - for example in GzipHeaderCodec.  We should patch this though regardless as right now its wrong should someone leverage the code in all cases.  This fixes that with no production hit as production calls are updated to specify the code directly.  Only test code now leverages the previously bugged constructor (now fixed) which now performs a hash (perf penalty).

Reviewed By: afrind

Differential Revision: D6746228

fbshipit-source-id: 7d3e2f74d08b1289bc3717a347c5ca70d1631649