Fix codec issue
Summary:
Apparently the codec can generate
more bytes that the input byte, makes sense.

However in the case where the

flow control limit is very close to the number
of pending bytes, the codec can generate bytes
into the write buffer which exceed the flow control
limit.

In this case, the txnEgressQueue doesn't have the
stream any more because the txn believes it has written
out all the bytes, however the bytes are technically only
partially written out. So the stream never gets a notification
to write out the remanining bytes and the eof gets stalled,
so the other side will never finish the stream.

Reviewed By: lnicco

Differential Revision: D6882204

fbshipit-source-id: 7d4d5f446c95237dbcb57e4ca1e0115ab753ffac