Be robust to newAcceptor() throwing exception
Summary:
If newAcceptor() throws an exception (e.g. if TLS certs / keys are not there),
the current code will throw an exception but not add an entry to workers_. Therefore, if
threadStopped() is called, it will find workers_ to be empty and fail the CHECK() call.

Instead, lets remove the CHECK() and add a comment explaining why this is not an invariant.

Differential Revision: D6985142

fbshipit-source-id: 37d3ee604b10c0beb8ac905dc7ed3a2a904be247