Expose a callback for HTTPServerAcceptor::onNewConnection.
Summary: This diff adds to HTTPServer the ability to hook a callback to run for every new connection (full handshake or resumption or no SSL). If the callback returns false, the socket will be reset instead of being further processed. This is useful for implementing client validation before the message is parsed.

Reviewed By: ngoyal

Differential Revision: D6986256

fbshipit-source-id: d21ec38f555badb1ad35b4f2c27dff9510ba4164