do not parse padding on CONTINUATION frames
Summary:
proxygen skips the padding or not, according to the padding flag in CONTINUATION frame.  I believe for most h2 implementations, the padding flag in CONTINUATION frame should be clean.  So it is a noop. But who knows what flags people will set it.

To be safer, just do not call parsePadding() and skipPadding() at all, for CONTINUATION frames.

Reviewed By: afrind

Differential Revision: D7195772

fbshipit-source-id: bd48748178f5d6815825fbfaae092b8196a3ef18