Remove unecessary mask inversion
Summary:
The bitmask in NBIT_MASKS was always inverted, so invert it in the constants file instead.  Further, the parameter names were always confusing since it used 'prefix' for something other than what the specification calls a prefix.  I changed it to 'instruction' and made it clear nbit is the length of the prefix.

Also fixed a bug in the old QCRAM implementation caused by the poor HPACKEncodeBuffer documentation: we used a 0 bit prefix (eg: reserve an entire byte for an instruction) instead of an 8 bit prefix (use the whole byte for integer).

Reviewed By: dddmello

Differential Revision: D7245521

fbshipit-source-id: 7fbcaeba0418926de92877dff080d04b8de62e97