Add THRIFT_CHANNEL_ID to SettingsId enum
Summary:
Thrift's H2Channel has used SettingId of value 100, which is not in the range of values of Experimental Use, that is between 0xf000 to 0xffff.

As SettingId of 100 is currently in use, a DEPRECATED version of this value along with THRIFT_CHANNEL_ID is added to the proxygen::SettingsId enum. After the services got updated, the THRIFT_CHANNEL_ID_DEPRECATED version will be deleted.

The kChannelSettingsId constexpr value is changed with these proxygen::SettingsId enum values.

Reviewed By: anirudhvr

Differential Revision: D7545528

fbshipit-source-id: 1063b0d1308cc1aa0720eb48bbb29a5cf044ec52