make associated stream ID to be optional
Summary:
We compare assocStreamId (HTTPCodec::StreamID) against the special value 0, to tell whether the transaction is pushed or not.  Then, encode the transaction with generateHeaders() or generatePushPromise() accordingly.

code is cleaner, by changing assocStreamId from HTTPCodec::StreamID to folly::Optional<HTTPCodec::StreamID>.  If assocStreamId has value, it must be a PushPromise.

from
  HTTPSession::createTransaction(
      StreamID streamID,
      StreamID assocStreamID = 0,
      folly::Optional<StreamID> controlStream,)

to

  HTTPSession::createTransaction(
      StreamID streamID,
      folly::Optional<StreamID> assocStreamID,
      folly::Optional<StreamID> controlStream,)

Reviewed By: afrind

Differential Revision: D7639511

fbshipit-source-id: 0daf3137a2d083bb3dd42ee2f369578f6215d5ff