Fix UBSAN warning in SPDYCodec::onGoaway
Summary:
The statusCode value is read in SPDYCodec::onControlFrame and is not validated to be a correct value.

We then try to cast it to an enum value in ::onGoAway, and that can be out of range.

Since the net result is to just return a protocol error, I default to that and do the conversion via the enum if in range.

Reviewed By: jas14

Differential Revision: D7714946

fbshipit-source-id: dcff935b0be6e07ef1a81b430b46fa03113ec5c3