Remove old "QPACK" implementation and refactor to prepare for new one
Summary:
This split is largely for ease of review.  The new QPACK shares almost none of the HPACK wire format beyond some primitives and has been mostly rewritten, so delete our 'built-in' QPACK for now.

This diff largely sets HPACK* classes back to their 2017-pre-QUIC compression experimentation state, but factors our the few bits that can be shared with QPACK.

Reviewed By: dddmello

Differential Revision: D7838748

fbshipit-source-id: 4e481fabed3cf3fc77b428a9da73a15c458d3ea6