QPACK implementation with current draft
Summary:
Current draft as of May 2018.

QPACKHeaderTable inherits from HeaderTable and adds logic and state required for QPACK:
  minUsable prevents referencing entries that will drain soon
  maxAcked indicates the acknowledged state of the decoder
  refCount tracks the per-entry outstanding reference count to prevent premature deletes

QPACKContext mirrors HPACKContext, but is actually quite thin.  Might be ok to delete it.

QPACKEncoder and QPACKDecoder mirror HPACKEncoder and HPACKDecoder, but inherit from QPACKContext instead of
HPACKContext.

QPACKCodec holds both a QPACKEncoder and QPACKDecoder.  Its mostly copypasta from HPACKCodec but the only ways out of that are templates or more allocations.

Reviewed By: lnicco

Differential Revision: D7838746

fbshipit-source-id: 1c1c20007665f931b9aeb0d81699543243a3530c