Do not send GOAWAY before SETTINGS
Summary: If the `notifyPendingShutdown` was called on the `HTTPSession` before `startNow` had been called then it was possible for a HTTP/2 connection to send a GOAWAY frame before the initial SETTINGS frame. This violates the spec and can blow up a client. Fix this by checking if we have started before sending GOAWAY and sending a GOAWAY after the initial SETTINGS if we are draining.

Reviewed By: afrind

Differential Revision: D7970911

fbshipit-source-id: d6fb35f5cf3ad22dcf96ccd9419ed16839ed99c1