Fix crash in HTTPServer::updateTLSCredentials
Summary:
HTTPServer::updateTLSCredentials crashes if there's an acceptor that isn't TLS. (My setup listens on 80 and 443, with only 443 being TLS.)

I modified the existing test so that it crashes without the change to HTTPServer.

(The fix could alternatively go in wangle::Acceptor::resetSSLContextConfigs, or wangle::SSLContextManager::resetSSLContextConfigs.)

Note that I also changed HTTPServer::updateTicketSeeds with a similar fix. Note that I haven't tested that fix or tested to see whether it's a problem.
Closes https://github.com/facebook/proxygen/pull/221

Reviewed By: w-o-o

Differential Revision: D8065590

Pulled By: udippant

fbshipit-source-id: 6d3a0c9f154a547e1791c06e2940c8d7eb01f64c