Lib changes to support websockets over h2
Summary:
Adds a field to http message for applications to request
a websocket upgrade. The codec encodes the request/response
when generating the buffer depending on the codec's direction.

The codec also indicates to the application when an upgrade request is
received or an upgrade happens on a previously sent request.

Reviewed By: afrind

Differential Revision: D8101569

fbshipit-source-id: c8b021c03b7ef87ad525847128334bee5127f5cf