remove encoder stream framing
Summary:
The consensus of the WG is to remove the length prefix for the encoder stream.  Therefore we need to support partial decodes.  If the buffer ends in the middle of an integer, that integer will be re-parsed each time more bytes arrive until the integer is complete.  If a literal length is parsed but the buf if shorter than the length, we discard the length and reparse it when more bytes arrive.  If we've parsed a literal name, we remember it and will re-start parsing from the value length.

This is somewhat suboptimal as we could remember the parsed lengths.  That would require changing the HPACKDecodeBuffer interface.  The worst case cost is that we reparse a literal's length once per byte of literal.  Room for improvement but this is a decent start.

Reviewed By: mjoras, w-o-o

Differential Revision: D8359270

fbshipit-source-id: 1a3b59921c5aa258d0a0356012b83c44d5842b29