Fix a DCHECK in HTTP2PriorityQueue
Summary:
Because the algorithm to correctly maintain the egress tree and enqueued weights is fairly complicated, this class uses a simple yet slower algorithm in debug mode to compare and ensure correctness.  There was a bug in the debug/check value calculation where it wasn't updated when a virtual node expired that was causing spurious check failures.

Note: in the course of debugging this, I found that adding children to a virtual node will cancel its expiration timer, but when a node converts from real to virtual (by having its transaction complete) it leaves the timer enabled.  This seems inconsistent but will fix in another diff

Reviewed By: lnicco

Differential Revision: D8924536

fbshipit-source-id: 6e6ed3a296c2f7db5883a6a32bcd28696a0e0d0c