tag implicit fallthroughs
Summary:
Generated code marked with `/*FALLTHROUGH*/` was failing with GCC on platform007 before. Now we use `sed` to replace the comments with FOLLY_FALLTHROUGH which works on all platforms.

I was tempted to also erase the `register` keywords but I can't guarantee that all our compilers are currently ignoring that.

Reviewed By: pixelb

Differential Revision: D8986133

fbshipit-source-id: 596dcc5b48644aa6033d06c7b2703e40a629b9bb