Don't add headers that would immediately drain
Summary: Any header that is immediately draining (eg: the header is larger than capacity - minFree) would never be used in compression.  Don't waste bytes on the wire by encoding them.  Note that because kMinFree is 48 and the minimum size of a header is 32, any tables smaller than 80 bytes can never hold a value.

Reviewed By: mjoras, lnicco

Differential Revision: D9819102

fbshipit-source-id: 0e81508b1040954088403168e7b79c2c9a227f53