Don't evict unacknowledged headers
Summary:
We had an issue in interop with non-blocking decoders and small tables.  The encoder would speculatively insert headers it *would* have indexed, but because there was no reference to it, there was nothing preventing the headers from being evicted by the encoder.  This would happen before the decoder acknowledged the insert, so there were lots of inserts but never any references to the table.

Now we require the header to be acknowledged before it becomes evictable.  This improved compression with very small tables (256 bytes) from -11% to 60%.

Reviewed By: mjoras

Differential Revision: D9819103

fbshipit-source-id: 0179e4bf330cb784087f1b37f579ceec6707f139