Simulate TSS for non-blocking decoders in immediate ack mode
Summary: When the simulator is using non-blocking mode with immediate acks, it is performing table inserts but not referencing those headers in header blocks.  The decoder must emit a Table State Sync instruction to tell the encoder the headers are available for use.  To do this correctly, we need to run the decoder during encoding.

Reviewed By: mjoras, lnicco

Differential Revision: D9819104

fbshipit-source-id: 925ace124d65dabf90012f6ddc6a9ab9c98af504