Don't name reference evicted entry
Summary: Edge case where inserting an entry to the table evicted the name entry I wanted to use.  From a compression standpoint, this solution is sub-optimal.  I could go back to the table and try to find a valid (non-draining, acknowledged) entry to name index.  But for now, just remove the name reference.

Reviewed By: mjoras

Differential Revision: D9931126

fbshipit-source-id: 4232c768317a27b289848cf5bf31e53a2838ce3e