hq -> hq-01, and make it easier to version
Summary: hq is reserved for the final version of HTTP over QUIC, which our implementation is a ways away from supporting.  The current draft is hq-15.  Since we're not quite there either, set the current version to hq-01 (no one speaks this).  We will revise the version every time we deploy a wire-incompatible change, until we reach full implementation of an official draft.

Reviewed By: lnicco

Differential Revision: D10341655

fbshipit-source-id: 9b71386042f024fa788fdc472756ccff8f4a14ee