Fix "explicitly assigning value to itself"
Summary:
A test in Proxygen triggers a new Clang warning that's been added to LLVM/Clang
trunk: `-Wself-assign-overloaded`. The test is explicitly testing this self
assignment, so it's not written in error. Add a pragma to ignore the new
warning for this test.

+cc compnerd since this is a case of two warnings being ignored in a row, and
I wanted his opinion on whether Clang should be patched to allow these two
ignores on the same line.

Reviewed By: taewookoh

Differential Revision: D10386924

fbshipit-source-id: 12b6828be8096fb99497b4dfaba1f78b36f5b0e3