Remove reset streams from queue
Summary: If a stream is reset while it has a block that is waiting to decode, we no longer need to decode it.  If we do, the cb pointer may be dangling.  This requires passing a stream ID into decodeStreaming.  There are a few other options - we could change the ownership model of the cb so that the decoder has partial ownership (shared_ptr), or we could return a handle via decodeStreaming which can be used for cancellation, but this felt like this most straightfoward.

Reviewed By: lnicco

Differential Revision: D10511204

fbshipit-source-id: b7a45b423083d595e225ff9833e98fadb175bf5c