setMaxTableSize also changes table capacity
Summary: If an encoder violates the protocol because it thinks the table is larger than the decoder's, it will result in an INVALID_INDEX reference when trying to decode the evicted header.  setMaxTableSize previously only set the maximum allowed value in a Table Size Update instruction.  Now it will also update the decoder's table size.

Reviewed By: mjoras

Differential Revision: D12910777

fbshipit-source-id: 3be303b935874828fab4c7079c320d4075d9b822