Prevent deadlock on encoder stream
Summary:
There's a nasty potential deadlock in HQ, where encoding a header block produces more output on the control stream than is available in its stream flow control window.  Depending on the egress pattern and amount of connection flow control remaining, the peer may receive a header block but never receive the control data it depends on.

This diff tells the encoder the maximum amount of control bytes it can generate that are referenceable by the stream block.  Once flow control is exhausted, the encoder will not insert any new values.  It will still process duplicates normally, but won't reference a duplicate if it didn't fit.

Reviewed By: lnicco

Differential Revision: D12903121

fbshipit-source-id: 8f1c896f2340d1a7940a20b6d95c1513ad7fc8a6