Delay applying decoder max table size until SETTINGS-ACK
Summary:
We were always cheating a little with this part of the spec.  The sender of the SETTINGS is not supposed to apply them until it receives the ACK.  So it's possible that a peer could have sent some headers encoded with the default (4k) header table, while a SETTINGS frame lowering the value was in flight.  This decode could have failed.

This is only the beginning or a partial fix for the issue.

1. It only applies to one setting, the dynamic table size.
2. It doesn't support more than one SETTING frame in flight
3. There is no timeout on the SETTINGS-ACK

Reviewed By: lnicco

Differential Revision: D12894342

fbshipit-source-id: f952209f4fa355d604b71c420a0e9552237b787b