Fix open source build with static folly
Summary:
Folly now only uses cmake, and doesn't build shared libraries by default.

As such, when proxygen attempts to use -lfolly, we must specify all of
folly's private dependencies
(which normally should be a call to pkg-config --static --libs folly)
afterwards.

We already use AC_CHECK_LIB for folly, so there is no need to explicitly
set it in LIBS.

This diff simply adds the remaining private folly dependencies in the correct
order so that the build will succeed.

This diff additionally does an AC_CHECK_LIB for sodium, which is a dependency
of Fizz.

Will probably switch to cmake soon, but this should be enough to fix the DSO
link errors when trying to compile with static folly.

Reviewed By: yfeldblum

Differential Revision: D13271108

fbshipit-source-id: 80f1c0f3c7ea5896cdba8c7ba0a11e0dc896a82e