isExpectingWindowUpdate -> isExpectingIngress
Summary:
We've observed cases with extremely stale (several days old) sessions hanging around. This appears to be caused by the fact that our send window is full and ingress was paused. The current logic for `isExpectingIngress`seems somewhat backwards, and in this situation causes us to never schedule the transaction timeout if our ingress is paused. This changes is to make it so that if we are expecting a window update then we are expecting ingress. This should prevent us from never timing out the transaction in the case that we never receive a window update.

Note that while this should address the problem of sessions growing out of control, it doesn't explain why we get so many sessions that end up like this in the first place.

Reviewed By: afrind

Differential Revision: D13466182

fbshipit-source-id: b60f3b79ef0b7733c947d45eac336068b03bb266