Fix memory leak in Base64 encode with openssl 1.1.0.
Summary:
Previously we set BIO_NOCLOSE so that the underlying BUF_MEM would not
get freed when the BIO was freed, and we manually freed the BUF_MEM. However in
openssl 1.1.0 there is an additional structure in the BIO that also doesn't get
freed if you set BIO_NOCLOSE, which leaked. Since we copy the data we want into
a std::string before freeing the BIO anyway, managing the BUF_MEM separately is
unnecessary.

Reviewed By: siyengar

Differential Revision: D13969881

fbshipit-source-id: 33dbe0029c2f9ff9ed45d79b08df9507d299f966