MaxEntries must be immutable
Summary:
The old code was calculating MaxEntries based on the table's capacity which is just *wrong*.  In particular, until the first instruction on the encoder stream is processed, the decoder's capacty is zero.  It would fail to decompress any request with a non-zero largestRef/RequiredInsertCount before that.

Now, we just calculate MaxEntries from the max table size.

Made several other fixes, bringing the implementation closer to the spec:

1. The initial size of both tables is 0
2. Disallow changes of the max size, except from 0 to non-zero

Reviewed By: lnicco

Differential Revision: D13982543

fbshipit-source-id: 644eceb15f0f5838fd64dfe06f35dea31423b268