Capture session byte offset for first and last body bytes
Summary:
Adds support for capturing the session byte offset of the first and last bytes of a response body.

Protocols that support preemption and multiplexing (e.g., HTTP/2) may write multiple response body in parallel to the transport. Capturing the first and last body byte offsets enables examination of this multiplexing.

The difference between these two offsets is also useful for measuring throughput as it provides the total number of bytes transferred via transport between the time the first byte of the response was flushed (`TransportInfo::timeToFirstByte`) and when the ack was received for the last byte in the response (`TransportInfo::timeToLastBodyByteAck`).

Notes:
- Added new functions to `HTTPTransactionTransportCallback` callback interface instead of adding parameters to `firstByteFlushed` and `lastByteFlushed` to avoid breaking existing code. New functions are not pure virtual.
- `onEgressBodyFirstByte` and `onEgressBodyLastByte` are called in `HTTPSessionBase::commonEom` (via `HTTPSessionBase::handleLastByteEvents`) and in `ByteEventTracker::processByteEvents`. Only capturing / including the offsets when called by `processByteEvents`, as the offset has already been recorded in the `ByteEvent` objects.

Reviewed By: udippant

Differential Revision: D14731317

fbshipit-source-id: b66517817d46442bd81bf88b8edf250966efc56c