use fb_github_project_workdir() for folly
Summary:
Update the folly spec to use `fb_github_project_workdir()` rather than just
`github_project_workdir()`.

This shouldn't have any impact for external github-based builds, but for our
internal continuous integration builds this should make builds for downstream
projects pick up the correct local version of folly that matches the commit of
the downstream project that is being built.

This behavior was unintentionally changed to not use the `fb_*` version in
D10520960.

Differential Revision: D14969380

fbshipit-source-id: 000b10525821fb5e0312d1869198a68befa0bd91