Don't include HPACK overhead in uncompressed bytes
Summary: HPACKHeader::bytes() returns the size of the headers plus 32, for HPACK table size calculations, but was erroneously being returned from emit(), which should only return the number of bytes emitted to the applciation.  This would cause inflated compression ratios when reported by the decoder.  The encoder logic is correct.

Reviewed By: lnicco

Differential Revision: D14939699

fbshipit-source-id: 54692cd8fc3dffa5acf7b4ecee07ca5a9ea3bb73