Add ostream include for operator<<
Summary:
libc++ provides the declaration of `operator<<(ostream, string)` inside
`<string>` but places its definition in `<ostream>`. We were pulling in
the former but not the latter, leading to an undefined reference at link
time. Explicitly pull in `<ostream>` to provide the definition.

It's strange that libc++ allows for this to occur, so I filed
https://bugs.llvm.org/show_bug.cgi?id=41563.

Reviewed By: narissiam

Differential Revision: D15040674

fbshipit-source-id: ecad6cd131579b349d056aae16a2ff1bedd2722b