Framework for tracking when first/last response byte sent by NIC
Summary:
Today we record a timestamp when the first and last bytes are written to the kernel buffer by a call to `send()` in the `folly::AsyncSocket` transport layer [1]. Because HTTP/2 supports pipelining, the kernel send buffer may not be empty when the first/last byte of a response is written via `send()` and as a result the byte may spend significant time in the buffer (depends on the client's throughput, latency, and socket LOWAT configuration).

This diff adds the framework for tracking when the NIC actually sends / transmits (TX) the first and last byte of a response via kernel timestamps.

This is useful for:
- analyzing how long response bytes are waiting in the kernel send buffer for HTTP/2 connections, which may help determine if the `LOWAT` setting for the socket is optimal for HTTP/2
- determining throughput, as it enables the time in the kernel send buffer to be accounted for in the total duration and enables adjacent responses to be coalesced when there are no "gaps" in the send queue

[1] The timestamps are triggered via the following chain: `AsyncSocket::writeImpl` -> `HTTPSession::WriteSegment::writeSuccess` (callback implementation for `AsyncSocket::WriteCallback`) -> `ByteEventTracker::processByteEvents` -> `HTTPTransaction::onEgressBodyFirstByte`.

Reviewed By: siyengar

Differential Revision: D14998833

fbshipit-source-id: 42408d9c56a0419c60ecf1306b577d16ff4219ad