Change SSL{Server,Common}Options::k{CipherList,SignatureAlgorithms} to static constexpr function
Summary:
Taking the address of a static constexpr variable will generate undefined
references until C++17, when static constexpr variables are implicitly inline.

Turning this into a static constexpr function generates weak symbols for both
C++14 & C++17. If anything, if you don't care about this, this makes it easier
to add/remove ciphers since you no longer need to adjust the array size
in the definition in SSLOptions.cpp.

The original motivation for this change was to fix a build failure in proxygen
oss, where declaring a wangle::SSLContextConfig would lead to an undefined
symbol linker error.

Reviewed By: lnicco

Differential Revision: D15246869

fbshipit-source-id: 44c016b9eecdf999efbffdcd381c024a6ffb11ac