Reduce noisy logging when we get buffer underflows
Summary:
We log errors when we receive partial table update or partial duplicate instructions, even though we correctly handle these two cases.

Here's a short explanation of why our handling of these two cases is correct: In either case, we set the value of partial_.consumed to be the bytes consumed before we parse that instruction (https://fburl.com/z69a7s2s). If we get a buffer underflow, we only consume the bytes that were consumed prior to the instruction that led to the buffer underflow (https://fburl.com/ggd5pmrd). The value of partial_.consumed isn't modified within either of these two cases.

Reviewed By: afrind

Differential Revision: D15271146

fbshipit-source-id: b92b40871a75fc8afd74f80b7c1a484a4534d6d9