Fix mismatched length error on sendEOM()
Summary:
In D14849318 I moved the counter tracking egress body bytes from
`sendBody()` to `sendBodyNow()`, but the check that prints the mismatch error is
still hapenning in `sendEOM()` which can potentially run before we call
sendBodyNow() - hence the (false) error message.

The fix is to return `actualResponseLength_` increment to where it was. Also, changed the `trimDeferredEgressBody()` to dump everything we have buffered.

Reviewed By: afrind

Differential Revision: D15569673

fbshipit-source-id: 55235d0f23488695507c22449115fba4349a4d00