Preface generation
Summary:
This diff simplifies the preface generation for the different types of the unidirectional streams by moving the (otherwise duplicated) `generateStreamPreface` method to `HQStreamBase` parent class.

This is achieved by making the `HQStreamBase` aware of the optional type of the stream. Classes that are derived from `HQStreamBase` are required to initialize the base class with the type to be able to use `generateStreamPreface`.

Since the base now generates the preface directly, it now also has access to the read/write buffers of the streams. This does not seem to pose any design problem.

               +---------------------------+
               |HQEgressPushStream         |
               |:HQStreamBase(hq:..:PUSH)  |
               |                           |
               +---------------------------+
  +-----------------------+  |  +---------------------------+
  |HQControlStream(type)  |  |  |HQIngressPushStream        |
  | :HQStreamBase(type)   |  |  | :HQStreamBase(hq:..:PUSH) |
  |                       |  |  |                           |
  +-----------------------+  |  +---------------------------+
              |              |                |
              +--------------+----------------+
                             v
                 +-----------------------+
                 |HQStreamBase           |
                 |                       |
                 |::generatePreface      |
                 +-----------------------+

Reviewed By: lnicco

Differential Revision: D15112796

fbshipit-source-id: 4092349d7a57775b8c7b7bf04cc00b7769459637