Disable + drain timestamp events when reads are shutdown
Summary:
This change adds logic to disable and drain any pending timestamp events when reads are shutdown. These timestamps cannot be received after reads are shutdown, so we must disable / drain or connection shutdown will stall.

Shutdown of an HTTP session is delayed if there are any pending byte tracker events. This ensures that we wait until write events have completed, but it also causes us to wait for other events such as socket timestamps. For instance, if we are waiting for a timestamp of the last byte ack (TTLBA), shutdown of the HTTP session will be delayed until either the timestamp arrives or the tracker event timeout.

Because of how libevent and folly::AsyncSocket work, the read callback must be installed to receive socket error messages (which include socket timestamp messages). However, we remove the read callback (shutdown reads) at various points in a connection's lifecycle. For instance, if the session has an error, or if the codec indicates that the session is not reusable (e.g., an HTTP/1.1 request with `Connection: close` in the headers, or an H2 request with a GOAWAY). Because of this, shutdown of HTTP sessions can be stalled on timestamp events that will never arrive. This creates a large problem in cases with `Connection: close`, as the client uses the connection closing as a signal that the response is complete and ends up stalling.

Notes:
- The underlying libevent library does not have a specific event type / registration for the error messages used to communicate timestamps -- it reuses the READ and WRITE types (1). This makes it impossible to register exclusively for error messages and is the reason why we must remain registered for read events (e.g., having a read callback installed) to get timestamps.
- We could install a dummy read callback when shutting down reads instead of removing it entirely. This would allow us to still receive timestamps, but it also removes the back pressure on the connection that explicitly shutting down reads creates. We could do a two phase approach (install dummy callback, and if client is non-complaint, drain and shutdown entirely), but this doesn't seem to be worth the complexity at the moment.

(1) https://github.com/libevent/libevent/blob/33053cdd8a9e1a7330b47759dee9cb209bed8f1b/epoll.c#L489-L490

Reviewed By: afrind

Differential Revision: D15441935

fbshipit-source-id: ca357c732b86d78ff2758799395cbf92e097c792