fbcode_builder: getdeps: fboss: fix libmnl source
Summary:
The source we were downloading from (netfilter.org) should work, but
unfortunately, is donw. This is causing getdeps FBOSS builds to fail. For now,
download from a mirror which seems to be up.

If this becomes a frequent issue with this repository/others, we can consider a
different solution.

Also, not sure how lfs pointer edit was missed by previous commit, added here.

Reviewed By: phshaikh

Differential Revision: D16070052

fbshipit-source-id: dd9278254c77e1a5845ca331a9eeb36f2bf03071