Do not kill transaction on stale skip/reject body offset
Summary:
Do not kill transaction on stale skip/reject body offset. This can happen when both skip/reject and send body are in flight, and the peer receives a stale offset after just having sent a body piece.
In such cases, no reason to kill transaction, just print an error and carry on.

Reviewed By: afrind

Differential Revision: D15638639

fbshipit-source-id: 1180e68bac890d7d323ab51ddcf64b2b69d905aa