HTTPTransaction: allow skips with offsets below already
Summary:
We need to be able to "cancel" bytes that have already been sent; this
change is to allow HTTP transaction to do that.
This is only useful when the body bytes of higher offset were not commited to
the wire yet, e.g. bytes are still pending in transaction/QUIC egress buffers
or were sent byt QUIC but haven't been acked yet.
If the bytes were already acked skipBodyTo() is going to be a no-op (will
actually return an error to the caller).

Reviewed By: afrind

Differential Revision: D15707475

fbshipit-source-id: ff8089e65882928fbade63a37c07f059264f1669