Periodic scheduling of the pool's rebuildIndex
Summary:
As can be seen in the below strobeligh, rebuildIndex is very expensive and is scheduled a lot.
rebuildIndex is done to balance the pools. It is triggered once a server in a pool change its state. Servers in production change state all the time. The resource consumption of rebuildIndex might magnify itself once there is a sev going-on which impacts backend servers.
The optimization is simply to run rebuildIndex in interval say by default of 100ms.  The interval value should configurable.
https://fburl.com/strobelight/yz3ypov2

Reviewed By: mjoras

Differential Revision: D16127399

fbshipit-source-id: 61d1de79d50915633fa1a1ef646f8be02e297c7c