Cleaning up HQ 1/4 - introducing HQ params
Summary:
Following the interop, I discovered that adding a simple Host parameter reqiures telescoping method invocations.

This diff is a suggestion of using a "param builder" pattern instead.

All flags and options are captured in the `HQParams` instance. A reference to that instance is passed to the various components of `hq`

- Extended the `HQParams` to allow initialization via `std::initializer_list`

Reviewed By: lnicco

Differential Revision: D16403156

fbshipit-source-id: 5558dc70e3f0ed22517ac0bcea6d16664710912c