HQ client
Summary:
Refactoring the `HQClient` initialization to depend on `HQParams` instead of accessing the flags directly.

Note: the implementation of `main.cpp` in this diff is a mess. The **intended** implementation is the one in the diff above, which changes `HQServer` initialization to rely on `HQParams`.

In this diff, however, `main.cpp` needs to initialize the `HQServer` via `FLAGS_` and the `HQClient` via `HQParams`, otherwise the diff will not compile, and will hide potential issues.

Reviewed By: lnicco

Differential Revision: D16533851

fbshipit-source-id: 1da78d5bbaa95eb60ddc6622afe027d10715bb65