Split TransactionByteEvent into separate target
Summary: It's currently not possible for the `HTTPTransaction` class to interact with `ByteEvent` objects due to a dependency loop that is created if `HTTPTransaction.h` includes `ByteEvents.h`. This isn't resolvable with forward declarations. Resolving this loop by shifting the `TransactionByteEvent` class (which references `HTTPTransaction`) is shifted into a separate class.

Reviewed By: sharma95

Differential Revision: D16550297

fbshipit-source-id: f6c76b8b9494d54d8bc6388f39c81c212974336d