Drop (som|eom)EventProcessed from ByteEventTracker
Summary:
Byte trackers that perform TX and ACK timestamp tracking need to know the next byte requiring timestamping when `preSend` is called. `somEventProcessed` and `eomEventProcessed` enable these byte trackers to maintain a cache of the next `FIRST_BYTE` and `LAST_BYTE` events respectively. However, this approach makes it difficult to support timestamping for other arbitrary bytes / events as it requires careful synchronization between the base ByteEventTracker and the caches held in the derived trackers, and it makes logic in `preSend` increasingly complex.

This diff changes existing `preSend` implementations so that they no longer cache the next event. Instead, `preSend` implementations will simply search from the start of the byte events until they find the next event that requires timestamping (according to the tracker's respective configuration) or reach the end of the list of byte events. In practice this adds very little overhead given that most events inserted do require tracking and allows us to avoid a good amount of error-prone complexity.

Reviewed By: sharma95

Differential Revision: D16547451

fbshipit-source-id: bffbbd6b96f696fae681358172676f51270030a5